+++ src/sympc/protocol/__init__.py
@@ -1,8 +1,9 @@
 """Implemented Protocols."""
 from . import beaver
 from . import spdz
+from .aby3 import ABY3
 from .falcon import Falcon
 from .fss import FSS
 from .protocol import Protocol
 
-__all__ = ["beaver", "spdz", "Falcon", "FSS", "Protocol"]
+__all__ = ["ABY3", "beaver", "spdz", "Falcon", "FSS", "Protocol"]
