+++ src/sympc/protocol/aby3/__init__.py
@@ -1,0 +1,5 @@
+"""ABY3 Protocol."""
+
+from .aby3 import ABY3
+
+__all__ = ["ABY3"]
