+++ src/sympc/protocol/falcon/falcon.py
@@ -7,11 +7,9 @@ arXiv:2004.02229 [cs.CR]
 from typing import Any
 from typing import Dict
 from typing import List
-from typing import Tuple
 
 # third party
 import torch
-import torchcsprng as csprng
 
 from sympc.protocol.protocol import Protocol
 from sympc.session import Session
@@ -74,82 +72,6 @@ class Falcon(metaclass=Protocol):
         return True
 
     @staticmethod
-    def trunc1(
-        ptr_list: List["ReplicatedSharedTensor"], shape: torch.Size, session: Session
-    ) -> List["ReplicatedSharedTensor"]:
-        """Apply the trunc1 algorithm in ABY3 for preprocessing.
-
-        Args:
-            ptr_list (List[ReplicatedSharedTensor]): Tensor to truncate
-            shape(torch.Size) : shape of tensor values
-            session(Session) : session the tensor belong to
-
-        Returns:
-            List["ReplicatedSharedTensor"] : Truncated shares.
-        """
-        gen = csprng.create_random_device_generator()
-        rand_value = torch.empty(size=shape, dtype=session.tensor_type).random_(
-            generator=gen
-        )
-        base = session.config.encoder_base
-        precision = session.config.encoder_precision
-        scale = base ** precision
-        x1 = ptr_list[0].get_copy().shares[0]
-        x2, x3 = ptr_list[1].get_copy().shares
-        x1_trunc = x1 >> precision if base == 2 else x1 // scale
-        x_trunc = (x2 + x3) >> precision if base == 2 else (x2 + x3) // scale
-        shares = [x1_trunc, x_trunc - rand_value, rand_value]
-        ptr_list = ReplicatedSharedTensor.distribute_shares(shares, session)
-        return ptr_list
-
-    @staticmethod
-    def getTruncationPair(x: MPCTensor, session: Session) -> Tuple[MPCTensor]:
-        """Generates truncation pair for the given MPCTensor.
-
-        Args:
-            x (MPCTensor): input tensor
-            session (Session) : session of the input tensor.
-
-        Returns:
-            Tuple[MPCTensor]: generate truncation pair.
-        """
-        r: List = []
-        rPrime: List = []
-        for session_ptrs in session.session_ptrs:
-            rst = session_ptrs.prrs_generate_random_share(shape=x.shape)
-            rst = rst.resolve_pointer_type()
-            rPrime.append(rst)
-            r.append(rst)
-
-        r = Falcon.trunc1(r, x.shape, session)
-        r_mpc = MPCTensor(shares=r, session=session, shape=x.shape)
-        rPrime_mpc = MPCTensor(shares=rPrime, session=session, shape=x.shape)
-        return r_mpc, rPrime_mpc
-
-    @staticmethod
-    def truncate(x: MPCTensor, session: Session) -> MPCTensor:
-        """Truncates the MPCTensor by scale factor.
-
-        Args:
-            x (MPCTensor): input tensor
-            session (Session) : session of the input tensor.
-
-        Returns:
-            MPCTensor: truncated MPCTensor.
-        """
-        x.share_ptrs = Falcon.trunc1(x.share_ptrs, x.shape, session)
-
-        """r,rPrime = Falcon.getTruncationPair(x,session)
-        shares: List = []
-        scale =(session.config.encoder_base**session.config.encoder_precision)
-        op = getattr(operator,"sub")
-        x_rp = op(x,rPrime)
-        x_rp = x_rp.reconstruct()//scale
-        result = r+ x_rp"""
-
-        return x
-
-    @staticmethod
     def mul_master(x: MPCTensor, y: MPCTensor, session: Session) -> MPCTensor:
         """Master method for multiplication.
 
@@ -170,6 +92,8 @@ class Falcon(metaclass=Protocol):
             NotImplementedError: Raised when implementation not present
 
         """
+        from sympc.protocol import ABY3
+
         if len(session.parties) != 3:
             raise ValueError("Falcon requires 3 parties")
 
@@ -196,7 +120,7 @@ class Falcon(metaclass=Protocol):
             )
             result = MPCTensor(shares=reshared_shares, session=x.session)
             result.shape = MPCTensor._get_shape("mul", x.shape, y.shape)  # for prrs
-            result = Falcon.truncate(result, session)
+            result = ABY3.truncate(result, session)
 
         else:
             raise NotImplementedError(
